refactor(mobile): split ActivityGraph.tsx 696->196 LOC (Cluster A2)
WHY: ActivityGraph was a 696-line monolith mixing grid geometry, a Supabase
fetch, a detail modal, and styling - over the project's 400-LOC ceiling and
with the heatmap's correctness (intensity bucketing) untestable while wired to
React state and a ScrollView.

Extracted into a co-located activity-graph/ directory:
- activity-grid.ts  - pure grid math (toDateStr, computeIntensity, buildGrid,
                      gridToColumns, formatTokens); now directly testable
- constants.ts      - cell dims + session/cost color palettes + ActivityMode
- useActivityData.ts- Supabase fetch + auth pre-flight + per-day aggregation
- DayDetailModal.tsx- presentational bottom-sheet for a tapped cell
- styles.ts         - shared StyleSheet
- __tests__/activity-grid.test.ts - 16 tests pinning the bucketing thresholds

The orchestrator (196 LOC) re-exports ActivityGraph + ActivityGraphProps, so
the public API via components/index.ts is unchanged. The production-incident
auth pre-flight WHY comment moved verbatim into the hook it documents.

Verification: mobile typecheck + lint clean; full suite 1738 pass (+12 new),
a11y regression guard green (accessibilityRole added to extracted touchables).
